엔티티정리
