+++ suriname-backend/src/main/java/com/suriname/category/Category.java
@@ -1,26 +1,35 @@
 package com.suriname.category;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "category")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
 public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long categoryId;
 
-    private String name;
-
-    private Boolean isVisible;
-
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "parent_id")
     private Category parent;
+
+    @Column(nullable = false, length = 50)
+    private String name;
+
+    @Column(nullable = false)
+    private Boolean isVisible;
+
+    @Builder
+    public Category(Category parent, String name, Boolean isVisible) {
+        this.parent = parent;
+        this.name = name;
+        this.isVisible = isVisible;
+    }
 }
