--- suriname-backend/src/main/java/com/suriname/category/entity/Category.java
@@ -1,35 +1,0 @@
-package com.suriname.category.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "category")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class Category {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long categoryId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "parent_id")
-    private Category parent;
-
-    @Column(nullable = false, length = 50)
-    private String name;
-
-    @Column(nullable = false)
-    private Boolean isVisible;
-
-    @Builder
-    public Category(Category parent, String name, Boolean isVisible) {
-        this.parent = parent;
-        this.name = name;
-        this.isVisible = isVisible;
-    }
-}
