+++ suriname-backend/src/main/java/com/suriname/customer/CustomerController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -48,12 +47,6 @@ public class CustomerController {
         return ResponseEntity.ok(Map.of("status", 200, "data", customer));
     }
     
-    // 수정
-    @PutMapping("/{id}")
-    public ResponseEntity<?> update(@PathVariable Long id, @RequestBody CustomerRegisterDto dto) {
-        customerService.update(id, dto);
-        return ResponseEntity.ok(Map.of("status", 200, "message", "수정 완료"));
-    }
     
     // 삭제
     @DeleteMapping("/{id}")
