+++ suriname-backend/src/main/java/com/suriname/customer/CustomerRepository.java
@@ -6,5 +6,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
-    Page<Customer> findAllByStatus(CustomerStatus status, Pageable pageable);
+	Page<Customer> findAllByStatus(Customer.Status status, Pageable pageable);
+
 }
