+++ suriname-backend/src/main/java/com/suriname/customer/CustomerService.java
@@ -1,7 +1,5 @@
 package com.suriname.customer;
 
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
 import java.util.Map;
 
 import org.springframework.data.domain.Page;
@@ -29,84 +27,70 @@ public class CustomerService {
 
     @Transactional
     public Map<String, Long> registerCustomer(CustomerRegisterDto dto) {
-        
-    	// 1. 고객 저장
-        Customer customer = new Customer();
-        customer.setName(dto.getName());
-        customer.setPhone(dto.getPhone());
-        customer.setEmail(dto.getEmail());
-        customer.setAddress(dto.getAddress());
-        customer.setBirth(dto.getBirth());
-        customer.setStatus(CustomerStatus.ACTIVE);
-        customer.setCreatedAt(Timestamp.valueOf(LocalDateTime.now()));
-        customer.setUpdatedAt(Timestamp.valueOf(LocalDateTime.now()));
+
+        // 고객 저장
+        Customer customer = Customer.builder()
+                .name(dto.getName())
+                .phone(dto.getPhone())
+                .email(dto.getEmail())
+                .address(dto.getAddress())
+                .birth(dto.getBirth())
+                .build();
+
         customerRepository.save(customer);
 
-        // 2. 제품 저장 또는 조회
+        // 제품 저장 또는 조회
         Product product;
         if (dto.getProduct().getProductId() != null) {
             product = productRepository.findById(dto.getProduct().getProductId())
                     .orElseThrow(() -> new RuntimeException("제품 없음"));
         } else {
-            product = new Product();
-            product.setProductName(dto.getProduct().getProductName());
-            product.setProductBrand(dto.getProduct().getProductBrand());
-            product.setModelCode(dto.getProduct().getModelCode());
-            product.setSerialNumber(dto.getProduct().getSerialNumber());
-            product.setIsVisible(true);
-            product.setCreatedAt(LocalDateTime.now());
-            product.setUpdatedAt(LocalDateTime.now());
-
-            // 카테고리 연동
+            // 새 제품 등록
             Category category = categoryRepository.findByName(dto.getProduct().getCategoryName())
                     .orElseThrow(() -> new RuntimeException("카테고리 없음"));
-            product.setCategory(category);
+
+            product = Product.builder()
+                    .productName(dto.getProduct().getProductName())
+                    .productBrand(dto.getProduct().getProductBrand())
+                    .modelCode(dto.getProduct().getModelCode())
+                    .serialNumber(dto.getProduct().getSerialNumber())
+                    .isVisible(true)
+                    .category(category)
+                    .build();
 
             productRepository.save(product);
         }
 
-        // 3. 고객 보유 제품 연결
-        CustomerProduct cp = new CustomerProduct();
-        cp.setCustomer(customer);
-        cp.setProduct(product);
-        cp.setCreatedAt(Timestamp.valueOf(LocalDateTime.now()));
-        cp.setUpdatedAt(Timestamp.valueOf(LocalDateTime.now()));
+        // 고객-제품 연결 
+        CustomerProduct cp = CustomerProduct.builder()
+                .customer(customer)
+                .product(product)
+                .build();
+
         customerProductRepository.save(cp);
 
-        return Map.of("customerId", customer.getCustomerId(), "customerProductId", cp.getCustomerProductId());
+        return Map.of(
+                "customerId", customer.getCustomerId(),
+                "customerProductId", cp.getCustomerProductId()
+        );
     }
 
     public Page<Customer> getAll(Pageable pageable) {
-        return customerRepository.findAllByStatus(CustomerStatus.ACTIVE, pageable);
+        return customerRepository.findAllByStatus(Customer.Status.ACTIVE, pageable);
     }
-    
-    // 상세 조회
+
     public Customer getDetail(Long id) {
         return customerRepository.findById(id)
-                .filter(c -> c.getStatus() == CustomerStatus.ACTIVE)
+                .filter(c -> c.getStatus() == Customer.Status.ACTIVE)
                 .orElseThrow(() -> new RuntimeException("고객 없음"));
     }
 
-    // 수정
-    @Transactional
-    public void update(Long id, CustomerRegisterDto dto) {
-        Customer customer = getDetail(id);
-        customer.setName(dto.getName());
-        customer.setPhone(dto.getPhone());
-        customer.setEmail(dto.getEmail());
-        customer.setAddress(dto.getAddress());
-        customer.setBirth(dto.getBirth());
-        customer.setUpdatedAt(Timestamp.valueOf(LocalDateTime.now()));
-        customerRepository.save(customer);
-    }
 
-    // 소프트 삭제
+    // 삭제
     @Transactional
     public void softDelete(Long id) {
         Customer customer = getDetail(id);
-        customer.setStatus(CustomerStatus.INACTIVE);
-        customer.setUpdatedAt(Timestamp.valueOf(LocalDateTime.now()));
+        customer.markAsInactive();
         customerRepository.save(customer);
     }
 }
-
