--- suriname-backend/src/main/java/com/suriname/customer/entity/Customer.java
@@ -1,75 +1,0 @@
-package com.suriname.customer.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "customer")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class Customer {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long customerId;
-
-    @Column(nullable = false, length = 10)
-    private String name;
-
-    @Column(nullable = false, length = 100)
-    private String email;
-
-    @Column(nullable = false, length = 20)
-    private String phone;
-
-    @Column(nullable = false, length = 255)
-    private String address;
-
-    @Column(nullable = false)
-    private LocalDate birth;
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(nullable = false)
-    private LocalDateTime updatedAt;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, length = 10)
-    private Status status;
-
-    public enum Status {
-        ACTIVE, INACTIVE
-    }
-
-    @PrePersist
-    public void onCreate() {
-        this.createdAt = LocalDateTime.now();
-        this.updatedAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    public void onUpdate() {
-        this.updatedAt = LocalDateTime.now();
-    }
-
-    @Builder
-    public Customer(String name, String email, String phone, String address, LocalDate birth) {
-        this.name = name;
-        this.email = email;
-        this.phone = phone;
-        this.address = address;
-        this.birth = birth;
-        this.status = Status.ACTIVE;
-    }
-
-    public void markAsInactive() {
-        this.status = Status.INACTIVE;
-    }
-}
