--- suriname-backend/src/main/java/com/suriname/customerproduct/entity/CustomerProduct.java
@@ -1,53 +1,0 @@
-package com.suriname.customerproduct.entity;
-
-import com.suriname.customer.entity.Customer;
-import com.suriname.product.entity.Product;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "customer_product")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class CustomerProduct {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long customerProductId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "customer_id", nullable = false)
-    private Customer customer;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id", nullable = false)
-    private Product product;
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
-    @Column(nullable = false)
-    private LocalDateTime updatedAt;
-
-    @PrePersist
-    public void onCreate() {
-        this.createdAt = LocalDateTime.now();
-        this.updatedAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    public void onUpdate() {
-        this.updatedAt = LocalDateTime.now();
-    }
-
-    @Builder
-    public CustomerProduct(Customer customer, Product product) {
-        this.customer = customer;
-        this.product = product;
-    }
-}
