+++ suriname-backend/src/main/java/com/suriname/satisfaction/entity/Satisfaction.java
@@ -1,6 +1,6 @@
 package com.suriname.satisfaction.entity;
 
-import com.suriname.customer.entity.Customer;
+import com.suriname.customer.Customer;
 import com.suriname.request.entity.Request;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
